+++ PhamiJam-Desktop/phamijam/lib/widgets/profile_grid/tiles/liked_songs_tile.dart
@@ -1,5 +1,9 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:phamijam/providers/friends_provider.dart';
 import 'package:phamijam/services/liked_songs_service.dart';
+import 'package:phamijam/widgets/profile_grid/tiles/private_tile_placeholder.dart';
+import 'package:provider/provider.dart';
 
 class LikedSongsTile extends StatelessWidget {
   const LikedSongsTile({super.key, required this.profileUid});
@@ -8,6 +12,10 @@ class LikedSongsTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final isOwner = profileUid == FirebaseAuth.instance.currentUser?.uid;
+    final isFriend = context.watch<FriendsProvider>().isFriend(profileUid);
+    if (!isOwner && !isFriend) return const PrivateTilePlaceholder();
+
     final colorScheme = Theme.of(context).colorScheme;
     return FutureBuilder<List<Map<String, dynamic>>>(
       future: LikedSongsService.fetchAllForUid(profileUid),